appy tasks ownerships

+++ src/tasks/tasks.service.ts
@@ -5,35 +5,41 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { TaskRepository } from './task.repository';
 import { Task } from './task.entity';
 import { TaskStatus } from './task-status.enum';
+import { User } from 'src/auth/user.entity';
 
 @Injectable()
 export class TasksService {
 
   constructor(@InjectRepository(TaskRepository) private taskRepository : TaskRepository ){}
 
-  async getTasks(filterDto : GetTaskFilterDto): Promise<Task[]> {
-    return this.taskRepository.getTasks(filterDto);
+  async getTasks(
+    filterDto : GetTaskFilterDto,
+    user: User,
+  ): Promise<Task[]> {
+    return this.taskRepository.getTasks(filterDto, user);
   }
 
-  async getTaskById(id: number) : Promise<Task> {
-    const found = await this.taskRepository.findOne(id);
+  async getTaskById(id: number, user: User) : Promise<Task> {
+    // await this.taskRepository.findOne(id);
+    const found = await this.taskRepository.findOne({ where: { id, userId: user.id } }); // (*)
     if (!found) throw new NotFoundException();
     return found;
   }
 
-  async updateTask(id: number, status: TaskStatus) {
-    const task = await this.getTaskById(id);
+  async updateTask(id: number, status: TaskStatus, user: User) {
+    const task = await this.getTaskById(id, user);
     task.status = status;
     await task.save();
     return task;
   }
 
-  async createTask(createTaskdto : CreateTaskDto): Promise<Task> {
-    return this.taskRepository.createTask(createTaskdto);
+  async createTask(createTaskdto : CreateTaskDto, user: User): Promise<Task> {
+    return this.taskRepository.createTask(createTaskdto, user);
   }
 
-  async deleteTask(id: number) : Promise<void> {
-    const result = await this.taskRepository.delete(id);
+  async deleteTask(id: number, user: User) : Promise<void> {
+    // await this.taskRepository.delete(id);
+    const result = await this.taskRepository.delete({ id, userId: user.id }); // (*) Notice here does not have where 
     if (result.affected === 0) throw new NotFoundException();
   }
 }
